帖子详情页开发

+++ app/src/main/java/com/app/jetpack/model/Comment.kt
@@ -8,7 +8,7 @@ data class Comment(
     var commentId: Long = 0L,
     var userId: Long = 0L,
     var commentType: Int = 0,
-    var createTime: Int = 0,
+    var createTime: Long = 0L,
     var commentCount: Int = 0,
     var likeCount: Int = 0,
     var commentText: String? = "",
